requirements.txt
absl-py==2.1.0
adjustText==1.3.0
aiobotocore==2.25.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiosignal==1.4.0
alembic==1.17.1
amqp==5.3.1
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
asyncssh==2.21.1
atpublic==6.0.2
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bidict==0.23.1
billiard==4.2.2
binaryornot==0.4.4
bleach==6.1.0
blinker==1.9.0
boto3==1.40.61
botocore==1.40.61
cachetools==6.2.0
celery==5.5.3
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
configobj==5.0.9
contourpy==1.2.1
controller==0.1.0
cookiecutter==2.6.0
cookiecutter-data-science==2.3.0
cryptography==43.0.3
cycler==0.12.1
Cython==3.0.10
databricks-sdk==0.71.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
dpath==2.2.0
dulwich==0.24.2
dvc==3.63.0
dvc-data==3.16.12
dvc-gdrive==3.0.1
dvc-gs==3.0.2
dvc-http==2.32.0
dvc-objects==5.1.2
dvc-render==1.0.2
dvc-s3==3.2.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
et_xmlfile==2.0.0
executing==2.0.1
fastapi==0.120.3
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==24.3.25
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
funcy==2.0
gast==0.5.4
gcsfs==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
grpcio==1.64.1
gto==1.9.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.6
httplib2==0.31.0
httpx==0.27.2
hydra-core==1.3.2
idna==3.7
imageio==2.34.1
importlib_metadata==8.7.0
iniconfig==2.1.0
ipykernel==6.29.4
ipython==8.24.0
isoduration==20.11.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterthemes==0.20.0
keras==3.3.3
kiwisolver==1.4.5
kombu==5.5.4
lazy_loader==0.4
lesscpy==0.15.1
libclang==18.1.1
lxml==5.2.2
Mako==1.3.10
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mlflow==3.5.1
mlflow-skinny==3.5.1
mlflow-tracing==3.5.1
mpmath==1.3.0
multidict==6.7.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.3.1
omegaconf==2.3.0
opencv-python==4.9.0.80
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opt-einsum==3.3.0
optree==0.11.0
orjson==3.11.3
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
patsy==1.0.2
pillow==10.3.0
platformdirs==4.2.1
pluggy==1.6.0
ply==3.11
prometheus_client==0.21.0
prompt-toolkit==3.0.43
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==5.9.8
pure-eval==0.2.2
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.12.0
pycparser==2.22
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
pydot==4.0.1
PyDrive2==1.21.3
pygit2==1.18.2
Pygments==2.18.0
pygtrie==2.5.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.9.0
python-json-logger==2.0.7
python-slugify==8.0.4
python-socketio==5.11.2
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.10
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.20.0
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
s3fs==2025.9.0
s3transfer==0.14.0
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.0
scmrepo==3.5.2
seaborn==0.13.2
semver==3.0.4
Send2Trash==1.8.3
setuptools==70.1.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.44
sqlparse==0.5.3
sqltrie==0.11.2
stack-data==0.6.3
starlette==0.49.1
statsmodels==0.14.5
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2024.5.10
tinycss2==1.3.0
tomlkit==0.13.3
torch==2.4.1
torchvision==0.19.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typer==0.19.2
types-python-dateutil==2.9.0.20241003
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
vine==5.1.0
voluptuous==0.15.2
waitress==3.0.2
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.43.0
wordcloud==1.9.3
wrapt==1.16.0
wsproto==1.2.0
xgboost==3.1.1
yarl==1.22.0
zc.lockfile==4.0
zipp==3.23.0
